+++ src/allegro/src/win/utf8.h
@@ -1,4 +1,4 @@
-// Copyright (C) 2013 by David Capello
+// Copyright (C) 2013, 2015 by David Capello
 
 #ifndef ALLEGRO_WIN_UTF8_H
 #define ALLEGRO_WIN_UTF8_H
@@ -6,7 +6,7 @@
 static char* convert_widechar_to_utf8(const wchar_t* wstr)
 {
   char* u8str;
-  int wlen = wcslen(wstr);
+  int wlen = (int)wcslen(wstr);
   int u8len = WideCharToMultiByte(CP_UTF8, 0, wstr, wlen, NULL, 0, NULL, NULL);
   u8len++;
   u8str = _AL_MALLOC(u8len);
@@ -18,7 +18,7 @@ static char* convert_widechar_to_utf8(const wchar_t* wstr)
 static wchar_t* convert_utf8_to_widechar(const char* u8str)
 {
   wchar_t* wstr;
-  int u8len = strlen(u8str);
+  int u8len = (int)strlen(u8str);
   int wlen = MultiByteToWideChar(CP_UTF8, 0, u8str, u8len, NULL, 0);
   wlen++;
   wstr = _AL_MALLOC(wlen * sizeof(wchar_t));
